requirements.txt
google-genai
elevenlabs
pymongo
python-dotenv
opencv-python
numpy
requests
flask
motor
fastapi
uvicorn
fastapi==0.111.0
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic==2.7.1

# Auth
python-jose[cryptography]==3.3.0
httpx==0.27.0

# Google AI
google-generativeai==0.7.2
Pillow==10.3.0

# MongoDB — motor 3.6.0 requires pymongo 4.9.x
motor==3.6.0
pymongo==4.9.2

# Snowflake
snowflake-connector-python==3.10.1

# WebSockets
websockets